+++ Modules/Intent.Modules.AspNetCore.Controllers.Interop.EntityFrameworkCore/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.0.1
+
+- Update: Please install Intent.AspNetCore.Controllers.Dispatch.ServiceContract as this module is no longer in use. It can be safely removed.
+
 ### Version 4.0.0
 
 - New: Upgraded Templates to use new Builder Pattern paradigm.